added authentication for add and remove

+++ src/main/java/com/qa/choonz/rest/controller/PlaylistController.java
@@ -70,13 +70,22 @@ public class PlaylistController {
     }
     
     @PostMapping("/add/{playlistId}/{trackId}")
-    public ResponseEntity<PlaylistDTO> add(@PathVariable Long playlistId,@PathVariable Long trackId){
-    	return new ResponseEntity<PlaylistDTO>(this.service.addTrack(playlistId, trackId), HttpStatus.ACCEPTED);
+    public ResponseEntity<PlaylistDTO> add(@PathVariable Long playlistId,@PathVariable Long trackId, @RequestHeader("token") String token){
+        if (AuthUtils.validToken(token)) {
+        	return new ResponseEntity<PlaylistDTO>(this.service.addTrack(playlistId, trackId), HttpStatus.ACCEPTED);
+        } else {
+        	return new ResponseEntity<PlaylistDTO>(HttpStatus.UNAUTHORIZED);
+        }
+    	
     }
     
     @PostMapping("/remove/{playlistId}/{trackId}")
-    public ResponseEntity<PlaylistDTO> remove(@PathVariable Long playlistId,@PathVariable Long trackId){
-    	return new ResponseEntity<PlaylistDTO>(this.service.removeTrack(playlistId, trackId), HttpStatus.ACCEPTED);
+    public ResponseEntity<PlaylistDTO> remove(@PathVariable Long playlistId,@PathVariable Long trackId, @RequestHeader("token") String token){
+        if (AuthUtils.validToken(token)) {
+        	return new ResponseEntity<PlaylistDTO>(this.service.removeTrack(playlistId, trackId), HttpStatus.ACCEPTED);
+        } else {
+        	return new ResponseEntity<PlaylistDTO>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
 }
